requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dotenv==0.9.9
fastapi==0.116.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.71
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.4
load-dotenv==0.1.0
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.97.1
orjson==3.10.18
packaging==24.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
